Move typedefs that describe targets from furnace.c to target.h
These typedefs should be in a single header file to allow
other source files to easily include them.
