+++ common.h
@@ -29,3 +29,5 @@
 
 #define STR(X) STR_HELPER(X)
 #define STR_HELPER(X) #X
+
+#define BUF_SIZE 64
